change namde ActionFinished and ActionDenied to inifinitive form
